reorganizing and extending the riff writer (now only wavwriter)

+++ qt-dab-6.9/forms-v7/new-display/scopes/spectrogramdata.cpp
@@ -23,8 +23,8 @@
  */
 #include	"spectrogramdata.h"
 //
-	spectrogramData::spectrogramData (double *data, int left,
-	                                  int width, int height,
+	spectrogramData::spectrogramData (double *data, float left,
+	                                  float width, int height,
 	                                  int datawidth, double max):
 	                                                 QwtRasterData () {
 	this	-> data		= data;
